Consolidate test processors
